refactor: move core event and parameter handling to new package
Moved event handling, metadata, and parameter classes to a new `core` package for improved code organization. Updated namespaces and file structure accordingly to better reflect functional responsibilities.
