+++ ndimmatrix/src/commonMain/kotlin/com/ndmatrix/core/event/EventChain.kt
@@ -1,116 +1,29 @@
-@file:Suppress("UNCHECKED_CAST")
+package com.ndmatrix.core.event
 
-package com.ndmatrix.parameter
-
-import com.ndmatrix.parameter.CallMetadata.Companion.CallMetadataKey
+import com.ndmatrix.core.metadata.CallMetadata
+import com.ndmatrix.core.metadata.PostExecMetadata
+import com.ndmatrix.core.parameter.ParameterHolder
+import com.ndmatrix.core.event.AbstractEventHandler
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.cancel
 import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.MutableSharedFlow
-import kotlinx.coroutines.flow.SharedFlow
-import kotlinx.coroutines.flow.SharingStarted
-import kotlinx.coroutines.flow.asSharedFlow
-import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.merge
-import kotlinx.coroutines.flow.shareIn
 import kotlinx.coroutines.launch
 import kotlin.coroutines.CoroutineContext
-import kotlin.coroutines.coroutineContext
-import kotlin.reflect.KClass
-import kotlin.reflect.safeCast
-import kotlin.time.Duration
-import kotlin.time.measureTime
 import kotlin.uuid.ExperimentalUuidApi
 import kotlin.uuid.Uuid
-
-/**
- * Contains metadata for the current event execution, enabling tracking of
- * nested call hierarchies.
- *
- * @property parentId the UUID of the parent event, or null if this is the root event.
- * @property currentId the UUID of the current event within the execution chain.
- *
- * Stored in the [CoroutineContext] to propagate identifiers between coroutines.
- */
-@OptIn(ExperimentalUuidApi::class)
-class CallMetadata(
-    val parentId: Uuid? = null,
-    val currentId: Uuid = Uuid.random(),
-) : CoroutineContext.Element {
-    companion object {
-        /**
-         * Key for storing and retrieving [CallMetadata] from the [CoroutineContext].
-         */
-        val CallMetadataKey = object : CoroutineContext.Key<CallMetadata> {}
-    }
-
-    override val key: CoroutineContext.Key<*>
-        /**
-         * Returns the key associated with this context element.
-         */
-        get() = CallMetadataKey
-}
-
-/**
- * Abstract base class for event handlers, providing infrastructure
- * to emit child events along with metadata.
- *
- * @param E the type of input [Message.Event].
- * @param coroutineContext the coroutine context used for subscribing to output events.
- * @param messageType the key to determinate which event must be processed.
- *
- * @property events a _hot_ [kotlinx.coroutines.flow.SharedFlow] of child events without metadata.
- * @property rawEvents a [kotlinx.coroutines.flow.SharedFlow] of pairs (parentId, Message) for internal routing.
- */
-@OptIn(ExperimentalUuidApi::class)
-abstract class AbstractEventHandler<E : Message.Event>(
-    coroutineContext: CoroutineContext,
-    messageType: KClass<E>
-) : PostMetadataEventHandler<E>(messageType) {
-    private val coroutineScope: CoroutineScope = CoroutineScope(coroutineContext)
-    private val _events = MutableSharedFlow<Pair<Uuid, Message>>()
-
-    /**
-     * A flow of child events without metadata.
-     */
-    val events = _events
-        .map { it.second }
-        .shareIn(coroutineScope, SharingStarted.Eagerly)
-
-    /**
-     * A flow of child events paired with their parent ID.
-     */
-    val rawEvents = _events.asSharedFlow()
-
-    /**
-     * Emits a new event into the outgoing events flow.
-     *
-     * @param block a factory lambda that produces a [Message] instance.
-     * @throws IllegalStateException if [CallMetadata] is not present in the context.
-     */
-    protected suspend fun returnEvent(block: () -> Message) {
-        val metadata = coroutineContext[CallMetadataKey]
-            ?: error("CallMetadata not found in CoroutineContext")
-        val parentId = metadata.currentId
-
-        // Launch a new coroutine for emission to avoid blocking the handler
-        coroutineScope.launch {
-            _events.emit(parentId to block())
-        }
-    }
-}
 
 /**
  * Manages sequential or parallel execution of event handlers and parameter holders
  * as a tree or chain layout.
  *
  * @param E the type of the starting [Message.Event] that triggers the chain.
- * @param intentsHandlers a list of [ParameterHolder] instances handling [Message.Intent]s.
+ * @param intentsHandlers a list of [com.ndmatrix.core.parameter.ParameterHolder] instances handling [Message.Intent]s.
  * @param eventsSender a list of [AbstractEventHandler] instances generating child events.
  * @param coroutineContext the coroutine context for the entire chain execution.
- * @param isDebug flag to enable collection of [PostExecMetadata] for debugging purposes.
+ * @param isDebug flag to enable collection of [com.ndmatrix.core.metadata.PostExecMetadata] for debugging purposes.
  *
- * @property coroutineScope the internal [CoroutineScope] for launching chain coroutines.
+ * @property coroutineScope the internal [kotlinx.coroutines.CoroutineScope] for launching chain coroutines.
  */
 @OptIn(ExperimentalUuidApi::class)
 abstract class EventChain<E : Message.Event>(
@@ -161,7 +74,7 @@ abstract class EventChain<E : Message.Event>(
      * @param e the initial event to start the chain.
      */
     fun general(e: E) {
-        coroutineScope.launch(CallMetadata(null, Uuid.random())) {
+        coroutineScope.launch(CallMetadata(null, Uuid.Companion.random())) {
             dispatchEvent(e)
         }
     }
@@ -239,7 +152,7 @@ abstract class EventChain<E : Message.Event>(
         message: Message.Intent
     ) {
         intentsHandlers.forEach { holder ->
-            launch(CallMetadata(parentId, Uuid.random())) {
+            launch(CallMetadata(parentId, Uuid.Companion.random())) {
                 holder.process(message)
             }
         }
@@ -250,7 +163,7 @@ abstract class EventChain<E : Message.Event>(
         message: Message.Event
     ) {
         eventsSender.forEach { target ->
-            launch(CallMetadata(parentId, Uuid.random())) {
+            launch(CallMetadata(parentId, Uuid.Companion.random())) {
                 target.process(message)
             }
         }
