+++ ndimmatrix/src/commonMain/kotlin/com/ndmatrix/core/event/IntentHandler.kt
@@ -1,0 +1,17 @@
+package com.ndmatrix.core.event
+
+import com.ndmatrix.core.event.Message
+import kotlin.reflect.KClass
+
+/**
+ * Marker class for [EventHandler]s that process [Message.Intent] instances.
+ *
+ * Differentiates intent handlers from general event handlers.
+ *
+ * @param E the specific subtype of [Message.Intent] this handler processes.
+ */
+interface IntentHandler<E : Message.Intent> : EventHandler<E>
+
+abstract class PostMetadataIntentHandler<E : Message.Intent>(
+    messageType: KClass<E>
+): PostMetadataEventHandler<E>(messageType), IntentHandler<E>
